write tests for Numbers, Booleans, Dates, and boxed Objects
